Merging dev to main to deploy to frontend

See merge request idatt2106-v21-09/backend!27

+++ src/main/java/edu/ntnu/idatt2106_09/backend/dto/GroceryItemShoppinglistDto.java
@@ -5,10 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class GroceryItemShoppinglistDto extends GroceryItemDto {
-    private int amount;
+    private double amount;
 }
